+++ scripts/tests/linux-keyring.sh
@@ -1,0 +1,7 @@
+#!/usr/bin/env bash
+# Figured out from this: <https://github.com/hwchen/keyring-rs/blob/master/linux-test.sh>
+
+set -euo pipefail
+
+echo -n "test" | gnome-keyring-daemon --unlock --replace
+cargo test -p firezone-windows-client
